+++ packages/rune/lib/rune.ml
@@ -504,7 +504,6 @@ let jit = Jit.jit
 let jit2 = Jit.jit2
 let jit_step = Jit.jit_step
 let jit' = Jit.jit'
-let to_device = Jit.to_device
 let pmap = Jit.pmap
 let pmap2 = Jit.pmap2